+++ Libplanet.Tests/Blockchain/BlockChainTest.MineBlock.cs
@@ -37,7 +37,7 @@ namespace Libplanet.Tests.Blockchain
                     privateKey: signingKey);
             _blockChainMinTx.StageTransaction(lightTx);
 
-            Func<long, int> getMaxBlockBytes = _blockChain.Policy.GetMaxBlockBytes;
+            Func<long, long> getMaxBlockBytes = _blockChain.Policy.GetMaxBlockBytes;
             Assert.Equal(1, _blockChain.Count);
             AssertBencodexEqual((Text)$"{GenesisMiner.ToAddress()}", _blockChain.GetState(default));
 
@@ -45,7 +45,7 @@ namespace Libplanet.Tests.Blockchain
             Block<DumbAction> block = await _blockChain.MineBlock(minerA);
             Assert.True(_blockChain.ContainsBlock(block.Hash));
             Assert.Equal(2, _blockChain.Count);
-            Assert.True(block.BytesLength <= getMaxBlockBytes(block.Index));
+            Assert.True(block.MarshalBlock().EncodingLength <= getMaxBlockBytes(block.Index));
             AssertBencodexEqual(
                 (Text)$"{GenesisMiner.ToAddress()},{minerA.ToAddress()}",
                 _blockChain.GetState(default)
@@ -55,7 +55,9 @@ namespace Libplanet.Tests.Blockchain
             Block<DumbAction> anotherBlock = await _blockChain.MineBlock(minerB);
             Assert.True(_blockChain.ContainsBlock(anotherBlock.Hash));
             Assert.Equal(3, _blockChain.Count);
-            Assert.True(anotherBlock.BytesLength <= getMaxBlockBytes(anotherBlock.Index));
+            Assert.True(
+                anotherBlock.MarshalBlock().EncodingLength <= getMaxBlockBytes(anotherBlock.Index)
+            );
             AssertBencodexEqual(
                 (Text)$"{GenesisMiner.ToAddress()},{minerA.ToAddress()},{minerB.ToAddress()}",
                 _blockChain.GetState(default)
@@ -65,7 +67,7 @@ namespace Libplanet.Tests.Blockchain
                 await _blockChain.MineBlock(new PrivateKey(), append: false);
             Assert.False(_blockChain.ContainsBlock(block3.Hash));
             Assert.Equal(3, _blockChain.Count);
-            Assert.True(block3.BytesLength <= getMaxBlockBytes(block3.Index));
+            Assert.True(block3.MarshalBlock().EncodingLength <= getMaxBlockBytes(block3.Index));
             AssertBencodexEqual(
                 (Text)$"{GenesisMiner.ToAddress()},{minerA.ToAddress()},{minerB.ToAddress()}",
                 _blockChain.GetState(default)
@@ -96,14 +98,14 @@ namespace Libplanet.Tests.Blockchain
                 await _blockChain.MineBlock(new PrivateKey(), append: false);
             Assert.False(_blockChain.ContainsBlock(block4.Hash));
             _logger.Debug(
-                $"{nameof(block4)}.{nameof(block4.BytesLength)} = {0}",
-                block4.BytesLength
+                $"{nameof(block4)}: {0} bytes",
+                block4.MarshalBlock().EncodingLength
             );
             _logger.Debug(
                 $"{nameof(getMaxBlockBytes)}({nameof(block4)}.{nameof(block4.Index)}) = {0}",
                 getMaxBlockBytes(block4.Index)
             );
-            Assert.True(block4.BytesLength <= getMaxBlockBytes(block4.Index));
+            Assert.True(block4.MarshalBlock().EncodingLength <= getMaxBlockBytes(block4.Index));
             Assert.Equal(3, block4.Transactions.Count());
             AssertBencodexEqual(
                 (Text)$"{GenesisMiner.ToAddress()},{minerA.ToAddress()},{minerB.ToAddress()}",